requirements.txt
streamlit==1.28.1
Pillow==10.1.0
rembg==2.0.50
onnxruntime==1.16.3
numpy==1.24.3
